+++ src/lib/util/style.ts
@@ -1,4 +1,4 @@
-import { TextFormat } from '../model';
+import { RequiredTextFormat, TextFormat } from '../model';
 
 export const DEFAULT_FONT_FAMILY = 'Arial';
 export const DEFAULT_FONT_SIZE = 14;
@@ -6,6 +6,101 @@ export const DEFAULT_FONT_COLOR = 'black';
 export const DEFAULT_STROKE_COLOR = DEFAULT_FONT_COLOR;
 export const DEFAULT_STROKE_WIDTH = 0;
 export const DEFAULT_STROKE_JOIN = 'round';
+
+/**
+ * Shallow-merges `TextFormat` sources into `target` while ignoring source properties that are
+ *  `undefined`, but deep-merges the `underline` and `strikethrough` properties if they are objects
+ *  on both.
+ * @param sources `TextFormat` sources to merge. Falsy sources are ignored/skipped.
+ * @returns Merged fully-spec'ed text format. Underline and strikethrough `thickness` will be
+ *  `NaN` if they should be set to whatever the relative thickness should be. `offset` will be
+ *  `NaN` if they should be set according to the current font.
+ */
+const _formatMerge = (
+  ...sources: Array<TextFormat | null | undefined | false | 0 | ''>
+): RequiredTextFormat => {
+  const target: RequiredTextFormat = {
+    fontFamily: DEFAULT_FONT_FAMILY,
+    fontSize: DEFAULT_FONT_SIZE,
+    fontWeight: '400',
+    fontStyle: '',
+    fontVariant: '',
+    fontColor: DEFAULT_FONT_COLOR,
+    strokeColor: DEFAULT_STROKE_COLOR,
+    strokeWidth: DEFAULT_STROKE_WIDTH,
+    underline: {
+      color: DEFAULT_FONT_COLOR,
+      thickness: 0,
+      offset: 0,
+    },
+    strikethrough: {
+      color: DEFAULT_FONT_COLOR,
+      thickness: 0,
+      offset: 0,
+    },
+  };
+
+  sources.forEach((source) => {
+    if (source && typeof source === 'object' && !Array.isArray(source)) {
+      Object.entries(source).forEach(([sourceKey, sourceValue]) => {
+        if (sourceValue !== undefined) {
+          // deep-merge underline and strikethrough objects
+          if (sourceKey === 'underline' || sourceKey === 'strikethrough') {
+            if (typeof sourceValue === 'boolean') {
+              if (sourceValue) {
+                // NOTE: using empty string as TBD once we know font color
+                target[sourceKey].color = '';
+                // NOTE: using NaN as TBD once we know font size/family
+                target[sourceKey].thickness = NaN;
+                target[sourceKey].offset = NaN;
+              }
+            } else if (
+              sourceValue &&
+              typeof sourceValue === 'object' &&
+              !Array.isArray(sourceValue)
+            ) {
+              const underStrikeSource = sourceValue as Partial<
+                RequiredTextFormat[typeof sourceKey]
+              >;
+              const underStrikeSourceKeys = Object.keys(
+                underStrikeSource
+              ) as (keyof RequiredTextFormat['underline'])[];
+              underStrikeSourceKeys.forEach((k) => {
+                if (underStrikeSource[k] !== undefined) {
+                  (target[sourceKey] as Record<string, unknown>)[k] =
+                    underStrikeSource[k];
+                }
+              });
+              if (!underStrikeSourceKeys.includes('color')) {
+                // since source specified an object (which enables underline) but did not
+                //  specify a color, set it to empty string so we know we need to set it to the
+                //  general font color once we're done
+                target[sourceKey].color = '';
+              }
+              if (!underStrikeSourceKeys.includes('thickness')) {
+                // since source specified an object (which enables underline) but did not
+                //  specify a thickness, set it to NaN so we know we need to set it to the
+                //  relative thickness once we're done
+                target[sourceKey].thickness = NaN;
+              }
+              if (!underStrikeSourceKeys.includes('offset')) {
+                // since source specified an object (which enables underline) but did not
+                //  specify a offset, set it to NaN so we know we need to set it to the
+                //  font-specific offset once we're done
+                target[sourceKey].offset = NaN;
+              }
+            }
+            // else, `sourceValue` is not boolean and not object: ignore
+          } else {
+            (target as Record<string, unknown>)[sourceKey] = sourceValue;
+          }
+        }
+      });
+    }
+  });
+
+  return target;
+};
 
 /**
  * Generates a text format based on defaults and any provided overrides.
@@ -16,22 +111,84 @@ export const DEFAULT_STROKE_JOIN = 'round';
 export const getTextFormat = (
   format?: TextFormat,
   baseFormat?: TextFormat
-): Required<TextFormat> => {
-  return Object.assign(
-    {},
-    {
-      fontFamily: DEFAULT_FONT_FAMILY,
-      fontSize: DEFAULT_FONT_SIZE,
-      fontWeight: '400',
-      fontStyle: '',
-      fontVariant: '',
-      fontColor: DEFAULT_FONT_COLOR,
-      strokeColor: DEFAULT_STROKE_COLOR,
-      strokeWidth: DEFAULT_STROKE_WIDTH,
-    },
-    baseFormat,
-    format
-  );
+): RequiredTextFormat => {
+  const underStrikeProps: Array<'underline' | 'strikethrough'> = [
+    'underline',
+    'strikethrough',
+  ];
+
+  // create a "full" version of `format` and `baseFormat` which have `underline` and
+  //  `strikethrough` specified as objects only to make ensuing merges easier
+  const fullBaseFormat = {
+    ...baseFormat,
+  };
+  const fullFormat = {
+    ...format,
+  };
+  underStrikeProps.forEach((prop) => {
+    [fullBaseFormat, fullFormat].forEach((obj) => {
+      if (obj[prop] !== undefined) {
+        if (obj[prop] === true) {
+          obj[prop] = {
+            thickness: NaN,
+          };
+        } else if (obj[prop] === false) {
+          obj[prop] = {
+            thickness: 0,
+          };
+        } else {
+          // must be an object: clone it to protect the original from upcoming merges
+          obj[prop] = {
+            ...obj[prop],
+          };
+        }
+      }
+    });
+  });
+
+  const result = _formatMerge(fullBaseFormat, fullFormat);
+
+  // NOTE: thickness and offset are subjective to the font being used, and in empirical testing
+  //  thus far, it looks like 1px thickness is good at 24px font size, and then it's directly
+  //  proportionate to the font size from there; it also looks to be the same directly proportionate
+  //  tweak for the offset based on 24px font size (i.e. certain fonts will require certain offsets
+  //  at 24px size, and then the offset should grow/shrink from there with the same factor as the
+  //  thickness)
+  const fontSizeFactor = result.fontSize / 24;
+
+  underStrikeProps.forEach((prop) => {
+    if (result[prop].color === '') {
+      result[prop].color = result.fontColor;
+    }
+
+    if (Number.isNaN(result[prop].thickness)) {
+      result[prop].thickness = fontSizeFactor;
+    }
+
+    // ❗️ IMPORTANT: always set the offset based on 24px FONT SIZE as that is the "zero"; offset
+    //  should be larger/smaller directly proportionate to the font size from that basis
+    if (Number.isNaN(result[prop].offset)) {
+      if (
+        result.fontFamily.startsWith('Roboto') ||
+        result.fontFamily.startsWith('Verdana')
+      ) {
+        result[prop].offset = -2;
+      } else if (
+        result.fontFamily === 'Inter' ||
+        result.fontFamily.startsWith('Montserrat')
+      ) {
+        result[prop].offset = -1;
+      } else if (result.fontFamily.startsWith('Comic Sans')) {
+        result[prop].offset = -5;
+      } else {
+        result[prop].offset = 0;
+      }
+
+      result[prop].offset *= fontSizeFactor;
+    }
+  });
+
+  return result;
 };
 
 /**